+++ release_blog.sh
@@ -1,1 +1,2 @@
 dart script/release_blog.dart
+dart script/produce_html.dart